+++ lib/src/bootstrap/browser.dart
@@ -7,9 +7,6 @@ import "../runner/plugin/remote_platform_helpers.dart";
 import "../util/stack_trace_mapper.dart";
 
 /// Bootstraps a browser test to communicate with the test runner.
-///
-/// This should NOT be used directly, instead use the `test/pub_serve`
-/// transformer which will bootstrap your test and call this method.
 void internalBootstrapBrowserTest(Function getMain()) {
   var channel =
       serializeSuite(getMain, hidePrints: false, beforeLoad: () async {